microservices/vector_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
chromadb
openai